fixed circular dependency
